requirements.txt
numpy
mss
Pillow
pynput